Created VideoFaceList
Function created to go through a video's frames
and stop as soon as a face is detected, returning the face.

A function for telling if a video has at least one
detected face or not is also created.

It is found that if only videos guaranteed to have at least
1 face detected are used to create a VideoFaceList, DataBunch
is creation and mini-batch loading seem to be reasonably fast,
at least not minutes, and there is no warning about broken dataset.